requirements.txt
altgraph==0.17.4
backports.tarfile==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
getmac==0.9.5
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.4.1
Kivy==2.3.0
Kivy-Garden==0.1.5
macholib==1.16.3
more-itertools==10.5.0
packaging==24.1
pillow==11.0.0
psutil==6.0.0
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pystray==0.19.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.3.0
six==1.16.0
urllib3==2.2.3
websockets==13.1
zipp==3.20.2
